+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import {Navbar, NavbarMenu, Home, About, Skills, Footer} from './Main';
+import {Navbar, NavbarMenu, Home, About, Skills, Footer, Experiences } from './Main';
 
 const main = ReactDOM.createRoot(document.getElementById('main'));
 main.render(
@@ -11,6 +11,7 @@ main.render(
     <Home />
     <About />
     <Skills />
+    <Experiences/>
     <Footer />
   </React.StrictMode>
 );